Improve tests and rename to ToolsPanel
