+++ packages/block-editor/src/hooks/margin.js
@@ -40,7 +40,7 @@ export function hasMarginValue( props ) {
 
 /**
  * Resets the margin block support attributes. This can be used when disabling
- * the margin support controls for a block via a progressive discovery panel.
+ * the margin support controls for a block via a `ToolsPanel`.
  *
  * @param {Object} props               Block props.
  * @param {Object} props.attributes    Block's attributes.
