--- packages/components/src/progressive-disclosure-panel/index.js
@@ -1,2 +1,0 @@
-export { default as ProgressiveDisclosurePanel } from './progressive-disclosure-panel';
-export { default as ProgressiveDisclosurePanelItem } from './progressive-disclosure-panel-item';
