--- packages/components/src/progressive-disclosure-panel/test/index.js
@@ -1,257 +1,0 @@
-/**
- * External dependencies
- */
-import { render, screen, fireEvent } from '@testing-library/react';
-
-/**
- * Internal dependencies
- */
-import {
-	ProgressiveDisclosurePanel,
-	ProgressiveDisclosurePanelItem,
-} from '../';
-
-const resetAll = jest.fn();
-
-// Default props for the progressive disclosure panel.
-const defaultProps = {
-	header: 'Panel header',
-	label: 'Display options',
-	resetAll,
-};
-
-// Default props for an enabled control to be rendered within panel.
-const controlProps = {
-	attributes: { value: true },
-	hasValue: jest.fn().mockImplementation( () => {
-		return !! controlProps.attributes.value;
-	} ),
-	label: 'Example',
-	onDeselect: jest.fn().mockImplementation( () => {
-		controlProps.attributes.value = undefined;
-	} ),
-	onSelect: jest.fn(),
-};
-
-// Default props without a value for an alternate control to be rendered within
-// the panel.
-const altControlProps = {
-	attributes: { value: false },
-	hasValue: jest.fn().mockImplementation( () => {
-		return !! altControlProps.attributes.value;
-	} ),
-	label: 'Alt',
-	onDeselect: jest.fn(),
-	onSelect: jest.fn(),
-};
-
-// Attempts to find the progressive disclosure panel via its CSS class.
-const getPanel = ( container ) =>
-	container.querySelector( '.components-progressive-disclosure-panel' );
-
-// Renders a default progressive disclosure panel including children that are
-// not to be represented within the panel's menu.
-const renderPanel = () => {
-	return render(
-		<ProgressiveDisclosurePanel { ...defaultProps }>
-			{ false && <div>Hidden</div> }
-			<ProgressiveDisclosurePanelItem { ...controlProps }>
-				<div>Example control</div>
-			</ProgressiveDisclosurePanelItem>
-			<ProgressiveDisclosurePanelItem { ...altControlProps }>
-				<div>Alt control</div>
-			</ProgressiveDisclosurePanelItem>
-			<span>Visible</span>
-		</ProgressiveDisclosurePanel>
-	);
-};
-
-// Helper to find the menu button and simulate a user click.
-const openDropdownMenu = () => {
-	const menuButton = screen.getByLabelText( defaultProps.label );
-	fireEvent.click( menuButton );
-};
-
-// Opens dropdown then selects the menu item by label before simulating a click.
-const selectMenuItem = async ( label ) => {
-	openDropdownMenu();
-	const menuItem = await screen.findByText( label );
-	fireEvent.click( menuItem );
-};
-
-describe( 'ProgressiveDisclosurePanel', () => {
-	describe( 'basic rendering', () => {
-		it( 'should render panel', () => {
-			const { container } = renderPanel();
-
-			expect( getPanel( container ) ).toBeInTheDocument();
-		} );
-
-		it( 'should render non panel item child', () => {
-			renderPanel();
-
-			const nonPanelItem = screen.queryByText( 'Visible' );
-
-			expect( nonPanelItem ).toBeInTheDocument();
-		} );
-
-		it( 'should render panel item flagged as default control even without value', () => {
-			render(
-				<ProgressiveDisclosurePanel { ...defaultProps }>
-					<ProgressiveDisclosurePanelItem { ...controlProps }>
-						<div>Example control</div>
-					</ProgressiveDisclosurePanelItem>
-					<ProgressiveDisclosurePanelItem
-						{ ...altControlProps }
-						isShownByDefault={ true }
-					>
-						<div>Alt control</div>
-					</ProgressiveDisclosurePanelItem>
-				</ProgressiveDisclosurePanel>
-			);
-
-			const altControl = screen.getByText( 'Alt control' );
-
-			expect( altControl ).toBeInTheDocument();
-		} );
-
-		it( 'should not render panel menu when there are no progressive panel items', () => {
-			render(
-				<ProgressiveDisclosurePanel { ...defaultProps }>
-					{ false && (
-						<ProgressiveDisclosurePanelItem>
-							Should not show
-						</ProgressiveDisclosurePanelItem>
-					) }
-					{ false && (
-						<ProgressiveDisclosurePanelItem>
-							Not shown either
-						</ProgressiveDisclosurePanelItem>
-					) }
-					<span>Visible but insignificant</span>
-				</ProgressiveDisclosurePanel>
-			);
-
-			const menu = screen.queryByLabelText( defaultProps.label );
-			expect( menu ).not.toBeInTheDocument();
-		} );
-
-		it( 'should render panel menu when at least one panel item', () => {
-			renderPanel();
-
-			const menuButton = screen.getByLabelText( defaultProps.label );
-			expect( menuButton ).toBeInTheDocument();
-		} );
-
-		it( 'should render reset all item in menu', async () => {
-			renderPanel();
-			openDropdownMenu();
-
-			const resetAllItem = await screen.findByRole( 'menuitem' );
-
-			expect( resetAllItem ).toBeInTheDocument();
-		} );
-
-		it( 'should render panel menu items correctly', async () => {
-			renderPanel();
-			openDropdownMenu();
-
-			const menuItems = await screen.findAllByRole( 'menuitemcheckbox' );
-
-			expect( menuItems.length ).toEqual( 2 );
-			expect( menuItems[ 0 ] ).toHaveAttribute( 'aria-checked', 'true' );
-			expect( menuItems[ 1 ] ).toHaveAttribute( 'aria-checked', 'false' );
-		} );
-
-		it( 'should disable default control menu items when the control has no value', async () => {
-			render(
-				<ProgressiveDisclosurePanel { ...defaultProps }>
-					<ProgressiveDisclosurePanelItem
-						{ ...controlProps }
-						isShownByDefault={ true }
-					>
-						<div>Example control</div>
-					</ProgressiveDisclosurePanelItem>
-					<ProgressiveDisclosurePanelItem
-						{ ...altControlProps }
-						isShownByDefault={ true }
-					>
-						<div>Alt control</div>
-					</ProgressiveDisclosurePanelItem>
-				</ProgressiveDisclosurePanel>
-			);
-			openDropdownMenu();
-
-			const menuItems = await screen.findAllByRole( 'menuitemcheckbox' );
-
-			expect( menuItems.length ).toEqual( 2 );
-			expect( menuItems[ 0 ] ).not.toHaveAttribute( 'disabled' );
-			expect( menuItems[ 1 ] ).toHaveAttribute( 'disabled' );
-		} );
-
-		it( 'should render panel header', () => {
-			renderPanel();
-			const header = screen.getByText( defaultProps.header );
-
-			expect( header ).toBeInTheDocument();
-		} );
-	} );
-
-	describe( 'conditional rendering of panel items', () => {
-		it( 'should render panel item when it has a value', () => {
-			renderPanel();
-
-			const exampleControl = screen.getByText( 'Example control' );
-			const altControl = screen.queryByText( 'Alt control' );
-
-			expect( exampleControl ).toBeInTheDocument();
-			expect( altControl ).not.toBeInTheDocument();
-		} );
-
-		it( 'should render panel item when corresponding menu item is selected', async () => {
-			renderPanel();
-			await selectMenuItem( altControlProps.label );
-			const control = await screen.findByText( 'Alt control' );
-
-			expect( control ).toBeInTheDocument();
-		} );
-
-		it( 'should prevent panel item rendering when toggled off via menu item', async () => {
-			renderPanel();
-			await selectMenuItem( controlProps.label );
-			const control = screen.queryByText( 'Example control' );
-
-			expect( control ).not.toBeInTheDocument();
-		} );
-	} );
-
-	describe( 'callbacks on menu item selection', () => {
-		beforeEach( () => {
-			jest.clearAllMocks();
-			controlProps.attributes.value = true;
-		} );
-
-		it( 'should call onDeselect callback when menu item is toggled off', async () => {
-			renderPanel();
-			await selectMenuItem( controlProps.label );
-
-			expect( controlProps.onSelect ).not.toHaveBeenCalled();
-			expect( controlProps.onDeselect ).toHaveBeenCalledTimes( 1 );
-		} );
-
-		it( 'should call onSelect callback when menu item is toggled on', async () => {
-			renderPanel();
-			await selectMenuItem( altControlProps.label );
-
-			expect( altControlProps.onSelect ).toHaveBeenCalledTimes( 1 );
-			expect( altControlProps.onDeselect ).not.toHaveBeenCalled();
-		} );
-
-		it( 'should call resetAll callback when its menu item is selected', async () => {
-			renderPanel();
-			await selectMenuItem( 'Reset all' );
-
-			expect( resetAll ).toHaveBeenCalledTimes( 1 );
-		} );
-	} );
-} );
