+++ packages/components/src/style.scss
@@ -30,7 +30,7 @@
 @import "./panel/style.scss";
 @import "./placeholder/style.scss";
 @import "./popover/style.scss";
-@import "./progressive-disclosure-panel/style.scss";
+@import "./tools-panel/style.scss";
 @import "./radio-control/style.scss";
 @import "./resizable-box/style.scss";
 @import "./responsive-wrapper/style.scss";
