+++ packages/components/src/tools-panel/index.js
@@ -1,0 +1,2 @@
+export { default as ToolsPanel } from './tools-panel';
+export { default as ToolsPanelItem } from './tools-panel-item';
